pyproject.toml
[project]
name = "cinetech"
version = "0.1.0"
description = "Applicazione che consente agli utenti di ricevere consigli personalizzati sui film"
readme = "README.md"
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "fastapi[standard]",
  "loguru",
  "pydantic",
  "pydantic-settings",
  "ipykernel",
]
requires-python = ">= 3.10"

[dependency-groups]
dev = [
    "coverage",
    "pre-commit",
    "pytest",
    "ruff",
    "langchain",
    "langchain-groq",
    "python-dotenv",
    "streamlit",
    "uvicorn[standard]",
    "requests",
    "python-dotenv",
    "numpy",
    "sentence-transformers",
    "chromadb",
    "tqdm",
    "llama-index",
    "llama-index-vector-stores-chroma",
    "llama-index-embeddings-huggingface-api",
    "llama-index-llms-huggingface-api",
    "llama-index-llms-ollama",
    "transformers[torch]",
]

[project.scripts]
cinetech = "cinetech.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
